requirements.txt
fastapi
uvicorn
opencv-python
pillow
python-multipart
numpy
